# [xy] Improve perf by batch calculating timeseries charts.