# Summary
<!-- Brief summary of what your code does -->
Reduce the analysis calculator time by optimizing timeseries calculation.
Timeseries calculation was not efficient due to duplicate calculation on buckets splitting and dataframe filtering. This PR changes to batch calculating timeseries charts.


# Tests
<!-- How did you test your change? -->
tested locally on a feature set with 10k rows and 2 datetime columns.
Tested 
*  Original time (Use multithread): 4.879s
*  Not use multithread: 4.5s
*  Not convert dateime: 4.1s
*  Not calculate value counts: 3917ms
*  Use mean instead of sum / len: 3895ms
*  Batch calculate timeseries: 1503ms

<img width="1079" alt="image" src="https://user-images.githubusercontent.com/80284865/172568873-e6f2d259-f528-49f9-abff-a1fb36035910.png">


cc:
<!-- Optionally mention someone to let them know about this pull request -->